:recycle: refactor : move StompInterceptor to global.config
